Extract ring FA3 autograd into ring_attn.py, remove deep nesting from attn.py
